Move DatabaseConnection, update imports
