+++ src/main/java/client/Character.java
@@ -78,6 +78,7 @@ import constants.skills.Shadower;
 import constants.skills.Sniper;
 import constants.skills.ThunderBreaker;
 import constants.skills.Warrior;
+import database.DatabaseConnection;
 import net.packet.Packet;
 import net.server.PlayerBuffValueHolder;
 import net.server.PlayerCoolDownValueHolder;
@@ -149,7 +150,6 @@ import server.partyquest.MonsterCarnival;
 import server.partyquest.MonsterCarnivalParty;
 import server.partyquest.PartyQuest;
 import server.quest.Quest;
-import tools.DatabaseConnection;
 import tools.LongTool;
 import tools.PacketCreator;
 import tools.Pair;
