+++ src/main/java/client/Family.java
@@ -21,12 +21,12 @@
  */
 package client;
 
+import database.DatabaseConnection;
 import net.packet.Packet;
 import net.server.Server;
 import net.server.world.World;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import tools.DatabaseConnection;
 import tools.PacketCreator;
 import tools.Pair;
