+++ src/main/java/client/inventory/ItemFactory.java
@@ -20,7 +20,7 @@
  */
 package client.inventory;
 
-import tools.DatabaseConnection;
+import database.DatabaseConnection;
 import tools.Pair;
 
 import java.sql.Connection;
