+++ src/main/java/database/DatabaseConnection.java
@@ -1,4 +1,4 @@
-package tools;
+package database;
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
