+++ src/main/java/net/server/channel/handlers/PlayerLoggedinHandler.java
@@ -39,6 +39,7 @@ import client.inventory.Pet;
 import client.keybind.KeyBinding;
 import config.YamlConfig;
 import constants.game.GameConstants;
+import database.DatabaseConnection;
 import net.AbstractPacketHandler;
 import net.packet.InPacket;
 import net.server.PlayerBuffValueHolder;
@@ -59,7 +60,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scripting.event.EventInstanceManager;
 import server.life.MobSkill;
-import tools.DatabaseConnection;
 import tools.PacketCreator;
 import tools.Pair;
 import tools.packets.WeddingPackets;
