+++ src/main/java/net/server/channel/handlers/RingActionHandler.java
@@ -30,6 +30,7 @@ import client.inventory.Item;
 import client.inventory.manipulator.InventoryManipulator;
 import client.processor.npc.DueyProcessor;
 import constants.id.ItemId;
+import database.DatabaseConnection;
 import net.AbstractPacketHandler;
 import net.packet.InPacket;
 import net.server.channel.Channel;
@@ -39,7 +40,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scripting.event.EventInstanceManager;
 import server.ItemInformationProvider;
-import tools.DatabaseConnection;
 import tools.PacketCreator;
 import tools.Pair;
 import tools.packets.WeddingPackets;
