+++ src/main/java/net/server/coordinator/session/SessionDAO.java
@@ -2,7 +2,7 @@ package net.server.coordinator.session;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import tools.DatabaseConnection;
+import database.DatabaseConnection;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
