+++ src/main/java/net/server/handlers/login/DeleteCharHandler.java
@@ -23,12 +23,12 @@ package net.server.handlers.login;
 
 import client.Client;
 import client.Family;
+import database.DatabaseConnection;
 import net.AbstractPacketHandler;
 import net.packet.InPacket;
 import net.server.Server;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import tools.DatabaseConnection;
 import tools.PacketCreator;
 
 import java.sql.Connection;
