+++ src/main/java/server/ExpLogger.java
@@ -1,7 +1,7 @@
 package server;
 
 import config.YamlConfig;
-import tools.DatabaseConnection;
+import database.DatabaseConnection;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
