+++ src/main/java/server/ItemInformationProvider.java
@@ -39,6 +39,7 @@ import constants.inventory.ItemConstants;
 import constants.skills.Assassin;
 import constants.skills.Gunslinger;
 import constants.skills.NightWalker;
+import database.DatabaseConnection;
 import net.server.Server;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -52,7 +53,6 @@ import provider.wz.WZFiles;
 import server.MakerItemFactory.MakerItemCreateEntry;
 import server.life.LifeFactory;
 import server.life.MonsterInformationProvider;
-import tools.DatabaseConnection;
 import tools.PacketCreator;
 import tools.Pair;
 import tools.Randomizer;
