+++ src/main/java/tools/mapletools/SimpleDatabaseConnection.java
@@ -1,9 +1,9 @@
 package tools.mapletools;
 
+import database.DatabaseConnection;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.core.config.Configurator;
-import tools.DatabaseConnection;
 
 import java.sql.Connection;
 import java.sql.SQLException;
